API upgrade
Better handling for GET request of Users and Posts
Packages upgrade
JSON now contains count of items
Common method that return formated JSON object when request is wrong or response is empty
Documentation upgrade
