+++ src/main/java/com/example/posting/app/PostingApplication.java
@@ -1,0 +1,3 @@
+package com.example.posting.app;
+
+// app path ?
